pyproject.toml
[project]
name = "rkc-cli"
version = "0.1.0"
description = "KT Odoo UI Test"
authors = [
    {name = "K-Tech", email = "info@k-tech.it"},
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1.8",
    "jinja2>=3.1.5",
    "python-slugify>=8.0.4",
    "selenium>=4.27.1",
    "webdriver-manager>=4.0.2",
]

[project.scripts]
rkc-cli = "rkc_cli.cli:main"

[tool.uv]

[dependency-groups]
dev = [
    "bumpversion>=0.6.0",
    "cairosvg>=2.7.1",
    "markupsafe>=3.0.2",
    "mdx-gh-links>=0.4",
    "mkdocs>=1.6.1",
    "mkdocs-autolinks-plugin>=0.7.1",
    "mkdocs-awesome-pages-plugin>=2.10.1",
    "mkdocs-click>=0.8.1",
    "mkdocs-embed-external-markdown>=3.0.2",
    "mkdocs-ezglossary-plugin>=1.7.1",
    "mkdocs-gen-files>=0.5.0",
    "mkdocs-get-deps>=0.2.0",
    "mkdocs-include-markdown-plugin>=7.1.2",
    "mkdocs-link-marker>=0.1.3",
    "mkdocs-macros-plugin>=1.3.7",
    "mkdocs-material>=9.5.49",
    "mkdocs-minify-plugin>=0.8.0",
    "mkdocs-redirects>=1.2.2",
    "mkdocs-simple-hooks>=0.1.5",
    "mkdocstrings[python]>=0.27.0",
    "pre-commit>=4.0.1",
    "pymarkdown>=0.1.4",
    "pymdown-extensions>=10.14",
    "pytest>=8.3.4",
]

[tool.ruff]
line-length = 120
indent-width = 4

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
